Move Radio to flat structure
