+++ packages/client/src/user/user-audit.ts
@@ -11,7 +11,7 @@
  */
 import { IFormField, RADIO_GROUP, TEXTAREA } from '@client/forms'
 import { messages } from '@client/i18n/messages/views/sysAdmin'
-import { RadioSize } from '@opencrvs/components/lib/forms'
+import { RadioSize } from '@opencrvs/components/lib/Radio'
 import { conditionals } from '@client/forms/utils'
 
 export interface IUserAuditForm {
