+++ packages/components/src/components/Radio/RadioGroup.tsx
@@ -10,9 +10,9 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 import * as React from 'react'
-import { RadioButton } from '../../interface/RadioButton'
-import { NoticeWrapper } from '../../DateField'
-import { InputLabel } from '../InputField/InputLabel'
+import { RadioButton } from '../interface/RadioButton'
+import { NoticeWrapper } from '../DateField'
+import { InputLabel } from '../forms/InputField/InputLabel'
 import styled from 'styled-components'
 
 const Wrapper = styled.div`
