+++ packages/components/src/components/Radio/index.ts
@@ -9,10 +9,5 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-export * from './InputField/InputField'
-export * from './InputField/InputError'
-export * from './TextInput'
-export * from './VerifyingIndicator'
-export * from './ValidIndicator'
-export * from './PDFViewer/PDFViewer'
-export * from './InputField/InputLabel'
+export * from './Radio'
+export * from './RadioGroup'
